Changing the visibility of inner Server condition classes (fixes #1483)
